Reorganized test folder structure.
