+++ app/client/src/tests/unit/!Banner.test.js
@@ -1,5 +1,5 @@
 test(`Write the banner`, () => {
     console.log(`##################################################
-#############  BEGIN TESTING HERE  ###############
+#############  BEGIN UNIT TESTING HERE  ###############
 ##################################################`);
 });
